+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # GFLabel 0.3.dev
 
+- Added `locknut` fragment.
 - Added `flanged` feature support to the `cullbolt` bolt style.
 - Added `wafer` head profile to the `cullbolt` bolt style.
 - Added `square_nut` fragment. Thanks to [@MinchinWeb](https://github.com/MinchinWeb).